requirements.txt
adal==1.2.7
applicationinsights==0.11.10
argcomplete==3.2.2
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-graphrbac==0.61.1
azure-identity==1.15.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azureml-automl-core==1.55.0.post1
azureml-core==1.55.0.post1
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.2
azureml-dataset-runtime==1.55.0
azureml-pipeline==1.55.0
azureml-pipeline-core==1.55.0
azureml-pipeline-steps==1.55.0
azureml-sdk==1.55.0
azureml-telemetry==1.55.0
azureml-train-automl-client==1.55.0
azureml-train-core==1.55.0
azureml-train-restclients-hyperdrive==1.55.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
contextlib2==21.6.0
cryptography==42.0.5
docker==7.0.0
fusepy==3.0.1
humanfriendly==10.0
idna==3.6
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.1
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
knack==0.11.0
msal==1.27.0
msal-extensions==1.1.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.23.5
oauthlib==3.2.2
packaging==23.2
paramiko==3.4.0
pathspec==0.12.1
pkginfo==1.9.6
portalocker==2.8.2
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
SecretStorage==3.3.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
